WIP try to refactor clone, realize this is going to be hard
Clone() expects a vector of Values today.  The vector can be different
in length if any of the values are optional.  This means it is ambiguous
to generate code to handle the old Clone() api.  We need to refactor
node() to make it so that all the things that interacted with Clone
the old way are updated to use fixed-length vectors of optionals instead
of variable length vectors of non-optionals.