Use `torch.empty()` instead of `torch.tensor()` in `torch.nn.Parameter`
Pull Request resolved: https://github.com/pytorch/pytorch/pull/66486

The newly-introduced Python dispatcher mode (`__torch_dispatch__`) does not have support for `torch.tensor()` (see #64360) and this causes friction in the user experience if some `nn.Modules` use `torch.tensor()` either implicitly or explicitly.

This PR replaces calls to `torch.tensor()` in `Parameter`, `UninitializedParameter`, and `UninitializedBuffer` with an equivalent call to `torch.empty()` which serves the same purpose and is syntactically more readable.
ghstack-source-id: 140520931

Differential Revision: [D31575587](https://our.internmc.facebook.com/intern/diff/D31575587/)