[5/n] Remove the Tensor subclass requirement for Python mode
With the `detach()` fix in #68954 the `__torch_dispatch__` type passed to `enable_python_mode()` now works for "vanilla" tensors (i.e. of type `torch.Tensor`). With that fix in place this PR relaxes the requirement for the argument passed to `enable_python_mode()` to be a Tensor subclass and allows *any* type with a `__torch_dispatch__` to be used in Python mode. Note that this is a backwards compatible change and does not affect the use of Tensor subclasses in Python mode.

Differential Revision: [D32688928](https://our.internmc.facebook.com/intern/diff/D32688928/)

ghstack-source-id: 144201280
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68965