In general, the current pattern allows for implementing optimisations for all the backends in a common place (see for example the optimisation for empty matrices).
Instantiate smaller U / V when calling cusolver_gesvdj in the cases when
full_matrices=False or compute_uv=False.

Now `linalg.svdvals` work as expected wrt the `out=` parameter.
Note that in the test `test_svd_memory_allocation` we were
passing a tensor of the wrong size and dtype and the test seemed to
pass...

Calls to cusolver_gesvd were incorrect for m <= n.