Implement _local_scalar_dense fallback for diagnostics
_local_scalar_dense is usually used by .item(), .any(), .all(),
methods on tensor to return a scalar summary for printing/logging.

Synchronous by nature, logging prints hurt lazy tensor performance and
should generally be avoided.  This diagnostic helps find the source.

In cases where _local_scalar_dense is used in part of a computation,
it would be beneficial to actually lower this operation, but it might
not be possible without API changes to make c10::Scalar behave lazily.