[functorch] refactor get_fallback_and_vmap_exhaustive
This PR refactors get_fallback_and_vmap_exhaustive into two helper
functions, `generate_vmap_inputs` and
`compute_quantities_for_vmap_test`. The former is responsible for
generating in_dims and the batched inputs; the latter is responsible for
computing vmap(blah) and the fallback.

The rationale for the refactor is that I want to use these pieces
separately in in-place vmap testing, which is coming soon.

Test Plan:
- run tests

ghstack-source-id: 2e1e73fd83e33acc5ee3d53c32acd5acc520e595
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82897