Update base for Update on "[functorch] Add vmap in-place testing"
After this PR
- test_vmap_exhaustive tests that both the inplace variant
and the out-of-place variant for an OpInfo is correct.
- test_op_has_batch_rule tests that both the inplace variant and the
out-of-place variant for an OpInfo have batching rules.

I added a separate denylist for failures in the inplace path for both of
these OpInfos. The separate denylist is not very nice, but I'll get rid
of it soon by fixing the problems (they're mostly low-hanging fruit)
in some future PRs.

Test Plan:
- wait for tests

[ghstack-poisoned]