Fix error_inputs for diag
`args` must be a tuple. It was not a tuple, so the error that was being
checked was actually being raised in the testing suite and not in
torch.diag itself.

This PR changes the error input to check something more reasonable: that
the input must be a matrix or a vector.

Test Plan:
- run tests

ghstack-source-id: fff440d44abd25d8cfd76b48bebcff54d0025a89
Pull Request resolved: https://github.com/pytorch/pytorch/pull/82984