Add mechanism to disable the "saved tensors hooks" feature
The rationale for this is that functorch doesn't work with saved
variable hooks at the moment or checkpointing and we need some way to
disable it.

Concretely:
- there's a context manager that does the disabling
- this feature is disabled on a thread-local basis
- one can set an error message or use the default error message that
says the feature has been disabled

Since it is thread local I needed to update ATen/ThreadLocalState. To
make things nicer, this PR refactors all the "saved tensors hooks"
related TLS things into a single struct.

Test Plan:
- new test

ghstack-source-id: 0621f3db45dd6508508350d8e915120857d5edef
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85553