Improve the performance of softmax in the mid range
We choose to go for a large block size. In any of the casees, the number of blocks per SM is 1,
so this increases the parallelism. Also, the PTX code seems to be more efficient
(the memory throughput goes from 69% to 77%). To go with this, we also increment the max RBLOCK.

ghstack-source-id: b61ac6ce550cb69873cc5833b9093c5087f3c4dc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/90994