address variable size unique_ptr in TensorImpl size checks
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96116

We move the size checks into the TensorImpl implementation file so
that they are only executed when compiling the TensorImpl class,
rather than anywhere the header is included. This neatly avoids the
problem when compiling it in a .cu file while also improving the
compile-time of the many compilation units that include TensorImpl.h.
ghstack-source-id: 181996436

Differential Revision: [D43834329](https://our.internmc.facebook.com/intern/diff/D43834329/)