Basic dynamo support for traceable allreduce
Attempt to make traceable allreduce work with torchdynamo,
bypassing problems with tracing the AsyncTensor subclass.

Accept a suboptimal solution for now, and optimize it later.
For now, wait happens immediately, which forces an early sync
in many cases.

Later, find a way either in dynamo or AOT stack to handle
AsyncCollectiveTensor to get the wait in the optimal place.

Note: It's actually not preferable to directly trace the current
traceable user api (e.g. _functional_collectives.all_reduce)
since it also registers the runtime value of tensor to work obj
in a map.  This registration should be done by backends instead.
- option 1: dynamo just special-cases these apis witout tracing
            and skips the reg (this PR)
- option 2: we actually trace these top-level apis,
            but perhaps dynamo special-cases the register_tensor api
            to be a no-op during tracing

ghstack-source-id: e2c0edf0ef4fcfac6f4ae9a4114cc7dd41078e04
Pull Request resolved: https://github.com/pytorch/pytorch/pull/94440