Fail fast when dynamo attempts to add unspecialized int/float as additional graph inputs
Pull Request resolved: https://github.com/pytorch/pytorch/pull/96786

Verified the changes to catch unspecialized int/floats being added as additional graph in D44037548 prior to RP(https://github.com/pytorch/pytorch/pull/95621).

However, with #95621 the issue to be solved originally is no longer valid because int & float in `forward` will always be specialized in export. This RP is to add the assertion anyway *(though not be hit unless there is a regression)* to immediately catch the attempt to add unspecialized int/float to additional graphargs

Differential Revision: [D44075910](https://our.internmc.facebook.com/intern/diff/D44075910/)

[ghstack-poisoned]