support statically linking c10 for tests
Summary:
The current c10 test build setup has the issue of requiring an export
for any symbol we wish to test. This can lead to exposing publicly
symbols we wish to remain private.

We can solve this problem by creating an object library for c10 named
c10.objlib, re-materializing it under the original name c10, and using
the objlib directly in tests.

Test Plan: Rely on CI.