[POC][pt2] Add canonicalizers to simplify expressions w/o decomposing
Currently decompositions serve a somewhat dual purpose. They either lower to
more primitive operators, or they inspect the arguments and specialize the
decomposition to a single simpler operator.

For example, `pow` specializes `pow(2, x)` into `exp2(x)` and `isinf`
short-circuits on integers to `zeros_like`.

These simplifications are great, but inductor can't always use them because they
are coupled to decomposition into prims. This either leads to duplication between
the decomposition and the lowering, or else one just gets forgotten.

This PR proposes adding a new registry for "canonicalizers" which only do
specialization and not decomposition. This means you can choose to use the
canonicalizer + decomposition, or just the canonicalizer on its own.

ghstack-source-id: e94de65a462f396c838698e067e77672957ea76f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/99003