Cleanup collectives and inductor backend.
Rename ForceInplace to InPlaceHint and use it with all collectives.
Split CollectiveKernel into InPlaceCollectiveKernel OutOfPlaceCollectiveKernel.

Split the lowering of inplace ops so that they always emit an InPlaceHint op
for their inputs. Return the InPlaceHint instances instead of the collective
op from `create`. This makes the collectives dead code, so we introduce
`has_side_effects` property to `BaseSchedulerNode`, override in
ExternKernelSchedulerNode and route it to the IR node, this prevents DCE
from killing the collective node.

Fix the wait behavior of all_reduce_coallesced so that you only need to
wait on a single tensor even when code-gen'ing. This requires ref-counting
the wait cleanup to ensure it won't trigger a wait too soon.

ghstack-source-id: 48b2152bea3afff9dbd1a5f2d4b9fd234eb98375
Pull Request resolved: https://github.com/pytorch/pytorch/pull/98641