Include yaml as PyTorch runtime dependency
The context behind this change is:
- We're adding a new custom operator API to PyTorch (#99439).
- This depends on `torchgen`, which is included as a part of PyTorch
- `torchgen` depends on the `yaml` project, which is not included as a
runtime dependency
- #99439 breaks the binary CI tests because it needs yaml

There are workarounds I can do, like refactor what I need from torchgen
to not import yaml. But the cleaner solution, because we do ship
torchgen with PyTorch, is to actually include yaml as a runtime
dependency.

Putting up this PR for discussion.

ghstack-source-id: 3bdf12f7ad04847b6d45a916c525c489a38421d2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/100166