[TensorBoard] Lazy initialize summary writers everywhere
Pull Request resolved: https://github.com/pytorch/pytorch/pull/100319

I've seen a lot of code that eagerly initializes a SummaryWriter instance with a logdir: which immediately creates a 40 byte file with a timestamp and file version, and then never uses it.

This is prevalent to the point that I could speed up our instance of TB significantly by simply skipping 40 byte files that had events in the name (some frameworks create them at 1 per rank); it seems worth adjusting this behavior globally for all PyTorch users.
ghstack-source-id: 188550061

Differential Revision: [D45425098](https://our.internmc.facebook.com/intern/diff/D45425098/)