[core][pruning] Add SemiStructuredSparse format to torch
Semi-Structured Sparsity, also known as 2:4 sparsity or fine-grained
structured sparsity, is a sparsity format introduced in late 2020 by
NVIDIA to it's Ampere GPUs.

This PR adds in a dummy format that describes the 2:4 sparsity layout,
which stores a sparse matrix by keeping track of the kept elements as
will as a mask stored in a metadata format.

ghstack-source-id: 113de92b8e15def2cd80c8f0e5c0051e30915ac7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102500