Implement bound_sympy
We implement a function that, given bounds on the values of the free
symbols of a sympy expression, it gives a bound on a the expression
itself.

We reshuffle a few things in value_ranges.py to create a
`SymPyValueRangeAnalysis` class, but we do not change any code really.
The only relevant change in that file is the addition of the
`sympy_bound`s function.

ghstack-source-id: 4e73c061751b1f81fe179639ca9a016a13d44b6b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/104559