[inductor] Add single pass "var_unnormalized" reduction_type
This is a bit inefficient because it computes the mean and throws it
away since ir.Reduction nodes only have 1 output. However, the mean
can at least be scheduled into the same loop as the variance now since
there is no data dependency. Thus we can take fewer passes over the
data.

ghstack-source-id: b4c68308576bbd4194724a233de2ccc5a0d8fe14
Pull Request resolved: https://github.com/pytorch/pytorch/pull/102486